requirements.txt
# Core PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# PyTorch Geometric and Graph Neural Networks
torch-geometric>=2.3.0
pyg-lib>=0.2.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0
torch-spline-conv>=1.2.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Materials Science and Crystal Structure
ase>=3.22.0
jarvis-tools>=2023.1.8

# Machine Learning and Training
ignite>=0.4.10
wandb>=0.15.0
tqdm>=4.64.0

# Data Storage and Processing
h5py>=3.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Development and Testing (Optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Distributed Training (Optional)
accelerate>=0.20.0

# Configuration and Utilities
pyyaml>=6.0
